ws: removes bufferutil class
