+++ lib/internal/websockets/Receiver.js
@@ -6,7 +6,6 @@
 
 const util = require('util');
 const BufferPool = require('internal/websockets/BufferPool');
-const BufferUtil = require('internal/websockets/BufferUtil');
 const PerMessageDeflate = require('internal/websockets/PerMessageDeflate');
 const opcodes = require('internal/websockets/Opcodes')
 
@@ -33,6 +32,32 @@ function fastCopy(length, srcBuffer, dstBuffer, dstOffset) {
     case 3: dstBuffer[dstOffset+2] = srcBuffer[2];
     case 2: dstBuffer[dstOffset+1] = srcBuffer[1];
     case 1: dstBuffer[dstOffset] = srcBuffer[0];
+  }
+}
+
+function bufferUtilUnmask(data, mask) {
+  var maskNum = mask.readUInt32LE(0, true);
+  var length = data.length;
+  var i = 0;
+  for (; i < length - 3; i += 4) {
+    var num = maskNum ^ data.readUInt32LE(i, true);
+    if (num < 0) num = 4294967296 + num;
+    data.writeUInt32LE(num, i, true);
+  }
+  switch (length % 4) {
+    case 3: data[i + 2] = data[i + 2] ^ mask[2];
+    case 2: data[i + 1] = data[i + 1] ^ mask[1];
+    case 1: data[i] = data[i] ^ mask[0];
+    case 0:;
+  }
+}
+
+bufferUtilMerge(mergedBuffer, buffers) {
+  var offset = 0;
+  for (var i = 0, l = buffers.length; i < l; ++i) {
+    var buf = buffers[i];
+    buf.copy(mergedBuffer, offset);
+    offset += buf.length;
   }
 }
 
@@ -331,7 +356,7 @@ Receiver.prototype.reset = function() {
  */
 
 Receiver.prototype.unmask = function (mask, buf, binary) {
-  if (mask != null && buf != null) BufferUtil.unmask(buf, mask);
+  if (mask != null && buf != null) bufferUtilUnmask(buf, mask);
   if (binary) return buf;
   return buf != null ? buf.toString('utf8') : '';
 };
@@ -346,7 +371,7 @@ Receiver.prototype.concatBuffers = function(buffers) {
   var length = 0;
   for (var i = 0, l = buffers.length; i < l; ++i) length += buffers[i].length;
   var mergedBuffer = new Buffer(length);
-  BufferUtil.merge(mergedBuffer, buffers);
+  bufferUtilMerge(mergedBuffer, buffers);
   return mergedBuffer;
 };
