+++ lib/internal/websockets/Sender.js
@@ -8,7 +8,6 @@ const events = require('events');
 const util = require('util');
 const EventEmitter = events.EventEmitter;
 const WSErrorCodes = require('websockets').WSErrorCodes;
-const BufferUtil = require('internal/websockets/BufferUtil');
 const PerMessageDeflate = require('internal/websockets/PerMessageDeflate');
 
 function writeUInt16BE(value, offset) {
@@ -42,6 +41,22 @@ function getRandomMask() {
     ~~(Math.random() * 255),
     ~~(Math.random() * 255)
   ]);
+}
+
+function bufferUtilMask(source, mask, output, offset, length) {
+  var maskNum = mask.readUInt32LE(0, true);
+  var i = 0;
+  for (; i < length - 3; i += 4) {
+    var num = maskNum ^ source.readUInt32LE(i, true);
+    if (num < 0) num = 4294967296 + num;
+    output.writeUInt32LE(num, offset + i, true);
+  }
+  switch (length % 4) {
+    case 3: output[offset + i + 2] = source[i + 2] ^ mask[2];
+    case 2: output[offset + i + 1] = source[i + 1] ^ mask[1];
+    case 1: output[offset + i] = source[i] ^ mask[0];
+    case 0:;
+  }
 }
 
 /**
@@ -231,7 +246,7 @@ Sender.prototype.frameAndSend = function(opcode, data, finalFragment, maskData, 
     outputBuffer[dataOffset - 2] = mask[2];
     outputBuffer[dataOffset - 1] = mask[3];
     if (mergeBuffers) {
-      BufferUtil.mask(data, mask, outputBuffer, dataOffset, dataLength);
+      bufferUtilMask(data, mask, outputBuffer, dataOffset, dataLength);
       try {
         this._socket.write(outputBuffer, 'binary', cb);
       }
@@ -241,7 +256,7 @@ Sender.prototype.frameAndSend = function(opcode, data, finalFragment, maskData, 
       }
     }
     else {
-      BufferUtil.mask(data, mask, data, 0, dataLength);
+      bufferUtilMask(data, mask, data, 0, dataLength);
       try {
         this._socket.write(outputBuffer, 'binary');
         this._socket.write(data, 'binary', cb);
